flask-server/requirements.txt
aioice
aiortc
alembic
aliyun-python-sdk-core
aliyun-python-sdk-dysmsapi
altgraph
amqp
APScheduler
av
bidict
billiard
blinker
cachelib
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
cryptography
distlib
dnspython
echarts
email_validator
eventlet
ffmpeg-python
filelock
Flask
Flask-Avatars
Flask-Caching
flask-cors
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Script
Flask-SocketIO
Flask-SQLAlchemy
Flask-WTF
future
gevent
google-crc32c
greenlet
gunicorn
h11
idna
ifaddr
itsdangerous
Jinja2
jmespath
jsonify
kombu
Mako
MarkupSafe
numpy
opencv-python
opencv-python-headless
packaging
pefile
peppercorn
pillow
platformdirs
prompt_toolkit
psutil
pycparser
pyee
pyinstaller
pyinstaller-hooks-contrib
pylibsrtp
PyMySQL
pyOpenSSL
python-dateutil
python-dotenv
python-engineio
python-socketio
redis
requests
setuptools
shortuuid
simple-websocket
six
SQLAlchemy
typing_extensions
tzdata
tzlocal
urllib3
vine
virtualenv
virtualenvwrapper-win
wcwidth
Werkzeug
wsproto
WTForms
zope.event
zope.interface